[#554] rolling back UserAccess class code reorganization
